+++ src/main/java/com/networknt/schema/JsonSchemaFactory.java
@@ -19,7 +19,8 @@ package com.networknt.schema;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.networknt.schema.SpecVersion.VersionFlag;
-import com.networknt.schema.uri.*;
+import com.networknt.schema.resource.*;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
