+++ src/main/java/com/networknt/schema/resource/MapSchemaMapper.java
@@ -1,4 +1,4 @@
-package com.networknt.schema.uri;
+package com.networknt.schema.resource;
 
 import java.util.Map;
 import java.util.function.Function;
