+++ src/main/java/com/networknt/schema/resource/SchemaLoader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 the original author or authors.
+ * Copyright (c) 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.networknt.schema.uri;
+package com.networknt.schema.resource;
 
-import java.io.IOException;
-import java.io.InputStream;
+import com.networknt.schema.AbsoluteIri;
 
 /**
- * InputStream source.
+ * Schema Loader used to load a schema given the retrieval IRI.
  */
 @FunctionalInterface
-public interface InputStreamSource {
+public interface SchemaLoader {
     /**
-     * Opens a new inputstream to the resource.
+     * Loads a schema given the retrieval IRI.
      * 
-     * @return a new inputstream
-     * @throws IOException
+     * @param absoluteIri the retrieval IRI
+     * @return the input stream source
      */
-    InputStream getInputStream() throws IOException;
+    InputStreamSource getSchema(AbsoluteIri absoluteIri);
 }
