+++ src/main/java/com/networknt/schema/resource/SchemaLoaders.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.networknt.schema.uri;
+package com.networknt.schema.resource;
 
 import java.util.ArrayList;
 import java.util.Collection;
