+++ src/main/java/com/networknt/schema/resource/UriSchemaLoader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.networknt.schema.uri;
+package com.networknt.schema.resource;
 
 import java.io.IOException;
 import java.io.InputStream;
