+++ src/test/java/com/networknt/schema/JsonSchemaFactoryUriCacheTest.java
@@ -2,8 +2,9 @@ package com.networknt.schema;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.networknt.schema.uri.InputStreamSource;
-import com.networknt.schema.uri.SchemaLoader;
+import com.networknt.schema.resource.InputStreamSource;
+import com.networknt.schema.resource.SchemaLoader;
+
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
